# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VendorHub - Multivendor Social Media, Sales & Inventory Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{id}/publish:
    post:
      operationId: publishPost
      tags: [posts]
      summary: Publish a post immediately
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  # ─── AI CONTENT ───────────────────────────────────────────────────────────
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── INVENTORY ────────────────────────────────────────────────────────────
...
  /sms-campaigns:
    get:
      operationId: listSmsCampaigns
      tags: [sms-campaigns]
      summary: List SMS campaigns
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of SMS campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmsCampaign"
    post:
      operationId: createSmsCampaign
      tags: [sms-campaigns]
      summary: Create an SMS campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsCampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
  /sms-campaigns/{id}:
    get:
      operationId: getSmsCampaign
      tags: [sms-campaigns]
      summary: Get an SMS campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SMS campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
    patch:
      operationId: updateSmsCampaign
      tags: [sms-campaigns]
      summary: Update an SMS campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsCampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmsCampaign"
  /sms-campaigns/{id}/send:
    post:
      operationId: sendSmsCampaign
      tags: [sms-campaigns]
      summary: Send an SMS campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignSendResult"

  # ─── ANALYTICS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Post:
      type: object
      required: [id, vendorId, caption, platforms, status, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        caption:
          type: string
        platforms:
          type: array
          items:
            type: string
        status:
          type: string
        mediaUrls:
          type: array
          items:
            type: string
        mediaType:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        hashtags:
          type: ["string", "null"]
        aiGenerated:
          type: boolean
        engagementData:
          type: ["string", "null"]
        createdAt:
          type: string
    PostInput:
      type: object
      required: [vendorId, caption, platforms]
      properties:
        vendorId:
          type: integer
        caption:
          type: string
        platforms:
          type: array
          items:
            type: string
        mediaUrls:
          type: array
          items:
            type: string
        mediaType:
          type: string
        scheduledAt:
          type: string
        hashtags:
          type: string
        aiGenerated:
          type: boolean
    PostUpdate:
      type: object
      properties:
        caption:
          type: string
        platforms:
          type: array
          items:
            type: string
        mediaUrls:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
        hashtags:
          type: string
        status:
          type: string

    # ─── AI ─────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      required: [id, vendorId, name, sku, price, stockQuantity, category, status, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        sku:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        costPrice:
          type: ["number", "null"]
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
        unit:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [vendorId, name, sku, price, stockQuantity, category]
      properties:
        vendorId:
          type: integer
        name:
          type: string
        sku:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        unit:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        category:
          type: string
        imageUrl:
          type: string
        status:
          type: string
        unit:
          type: string

    # ─── INVENTORY ──────────────────────────────────────────────────────────
...
    CampaignSendResult:
      type: object
      required: [sent, failed, message]
      properties:
        sent:
          type: integer
        failed:
          type: integer
        message:
          type: string

    # ─── SMS CAMPAIGN ───────────────────────────────────────────────────────
    SmsCampaign:
      type: object
      required: [id, vendorId, name, message, status, recipientCount, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        message:
          type: string
        status:
          type: string
        recipientCount:
          type: integer
        sentCount:
          type: integer
        targetAudience:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
    SmsCampaignInput:
      type: object
      required: [vendorId, name, message]
      properties:
        vendorId:
          type: integer
        name:
          type: string
        message:
          type: string
        targetAudience:
          type: string
        scheduledAt:
          type: string
        recipientCount:
          type: integer
    SmsCampaignUpdate:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        targetAudience:
          type: string
        scheduledAt:
          type: string
        status:
          type: string

    # ─── ANALYTICS ──────────────────────────────────────────────────────────
...