# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * VendorHub - Multivendor Social Media, Sales & Inventory Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  vendorId: number;
  name: string;
  sku: string;
  description?: string;
  price: number;
  costPrice?: number;
  stockQuantity: number;
  lowStockThreshold?: number;
  category: string;
  imageUrl?: string;
  unit?: string;
}